Add default layout; use CDNs for vendor assets
It is preferable to use CDNs or external sources in general for vendor assets.
Phoenix.js is part of the framework but as it's not bundled with it, it is
better to use a CDN than keeping the code with our custom code. This way it's
easier to maintain the whole app and make updates when required. The file
shouldn't be in our Git history at all.